Remove option to run custom write benchmark
The reasoning for this is that the default benchmarks should provide
adequate information regarding the performance of different event sizes,
batch sizes, and FoundationDB value sizes. So long as the benchmark
results are recorded, it's unlikely they'll need to be re-run.

The benchmarks are only for tuning purposes: the results will yield the
optimal values to set as macros in the code. In addition, the benchmarks
are being used to independently verify information readily available in
the FoundationDB docs.